+++ packages/expo-barcode-scanner/CHANGELOG.md
@@ -8,6 +8,7 @@
 
 - Converted plugin to TypeScript. ([#11715](https://github.com/expo/expo/pull/11715) by [@EvanBacon](https://github.com/EvanBacon))
 - Updated Android build configuration to target Android 11 (added support for Android SDK 30). ([#11647](https://github.com/expo/expo/pull/11647) by [@bbarthec](https://github.com/bbarthec))
+- Added support for string literal as bar code types input in iOS ([#00000](https://github.com/expo/expo/pull/00000) by [@jarvisluong](https://github.com/jarvisluong))
 
 ### 🐛 Bug fixes